Refactor new chart into separate components
