+++ apps/web/ui/charts/areas.tsx
@@ -1,0 +1,85 @@
+import { LinearGradient } from "@visx/gradient";
+import { Group } from "@visx/group";
+import { Area, AreaClosed, Circle } from "@visx/shape";
+import { motion } from "framer-motion";
+import { Fragment, useMemo } from "react";
+import { useChartContext, useChartTooltipContext } from "./chart-context";
+
+export default function Areas() {
+  const { data, series, margin, xScale, yScale } = useChartContext();
+  const { tooltipData } = useChartTooltipContext();
+
+  const zeroedData = useMemo(() => {
+    return data.map((d) => ({
+      ...d,
+      values: Object.fromEntries(Object.keys(d.values).map((key) => [key, 0])),
+    })) as typeof data;
+  }, [data]);
+
+  return (
+    <Group left={margin.left} top={margin.top}>
+      {series.map((s) => (
+        <Fragment key={s.id}>
+          <LinearGradient
+            className="text-blue-500"
+            id={`${s.id}-background`}
+            fromOffset="20%"
+            from="currentColor"
+            fromOpacity={0.01}
+            to="currentColor"
+            toOpacity={0.2}
+            x={0}
+            y1={1}
+          />
+
+          {/* Area */}
+          <AreaClosed
+            data={data}
+            x={(d) => xScale(d.date)}
+            y={(d) => yScale(s.accessorFn(d) ?? 0)}
+            yScale={yScale}
+          >
+            {({ path }) => {
+              return (
+                <motion.path
+                  initial={{ d: path(zeroedData) || "" }}
+                  animate={{ d: path(data) || "" }}
+                  fill={`url(#${s.id}-background)`}
+                />
+              );
+            }}
+          </AreaClosed>
+
+          {/* Line */}
+          <Area
+            data={data}
+            x={(d) => xScale(d.date)}
+            y={(d) => yScale(s.accessorFn(d) ?? 0)}
+          >
+            {({ path }) => (
+              <motion.path
+                initial={{ d: path(zeroedData) || "" }}
+                animate={{ d: path(data) || "" }}
+                className="text-blue-700"
+                stroke="currentColor"
+                strokeOpacity={0.8}
+                strokeWidth={2}
+              />
+            )}
+          </Area>
+
+          {/* Latest value dot */}
+          {!tooltipData && (
+            <Circle
+              cx={xScale(data.at(-1)!.date)}
+              cy={yScale(s.accessorFn(data.at(-1)!))}
+              r={4}
+              className="text-blue-700"
+              fill="currentColor"
+            />
+          )}
+        </Fragment>
+      ))}
+    </Group>
+  );
+}
